+++ src/master_eqn.c
@@ -1,6 +1,6 @@
 #include<include/master.h>
 
-Vec f(double t, Vec y){
+Vec f(long double t, Vec y){
     Vec function = vecInitZerosA(y.len);
     // *(vecRef(function, 0)) = *vecRef(y, 2) + *vecRef(y, 1) + 2;
     // *(vecRef(function, 1)) = *vecRef(y, 2) + *vecRef(y, 0) + 3;
@@ -11,16 +11,16 @@ Vec f(double t, Vec y){
     return function;
 }
 
-rk45 rkf45_calculator(double h, double t_i, Vec y_i){
+rk45 rkf45_calculator(long double h, long double t_i, Vec y_i){
     size_t l = y_i.len;
 
-    // double k_0 = h * f(t_i, y_i);
+    // long double k_0 = h * f(t_i, y_i);
 
     Vec k_0 = f(t_i, y_i); 
     vecScale(h, k_0, &k_0);
 
 
-    // double k_1 = h * f(t_i + h * 1 / 4, y_i + k_0 / 4);
+    // long double k_1 = h * f(t_i + h * 1 / 4, y_i + k_0 / 4);
 
     Vec vec11 = vecInitZerosA(l);
     vecScale(0.25, k_0, &vec11);
@@ -29,7 +29,7 @@ rk45 rkf45_calculator(double h, double t_i, Vec y_i){
     vecScale(h, f(t_i + h * 1.0 / 4.0, vec11), &k_1);
 
 
-    // double k_2 = h * f(t_i + h * 3 / 8, y_i + k_0 * 3 / 32 + k_1 * 9 / 32);
+    // long double k_2 = h * f(t_i + h * 3 / 8, y_i + k_0 * 3 / 32 + k_1 * 9 / 32);
 
     Vec vec21 = vecInitZerosA(l);
     vecScale(3.0 / 32.0, k_0, &vec21); 
@@ -41,7 +41,7 @@ rk45 rkf45_calculator(double h, double t_i, Vec y_i){
     vecScale(h, f(t_i + h * 3.0 / 8.0, vec21), &k_2);
 
 
-    // double k_3 = h * f(t_i + h * 12 / 13, y_i + k_0 * 1932 / 2197 + k_1 * (-7200) / 2197 + k_2 * 7296 / 2197);
+    // long double k_3 = h * f(t_i + h * 12 / 13, y_i + k_0 * 1932 / 2197 + k_1 * (-7200) / 2197 + k_2 * 7296 / 2197);
 
     Vec vec31 = vecInitZerosA(l);
     vecScale(1932.0 / 2197.0, k_0, &vec31); 
@@ -56,7 +56,7 @@ rk45 rkf45_calculator(double h, double t_i, Vec y_i){
     vecScale(h, f(t_i + h * 12.0 / 13.0, vec31), &k_3);
 
 
-    // double k_4 = h * f(t_i + h, y_i + k_0 * 439 / 216 + k_1 * (-8) + k_2 * 3680 / 513 + k_3 * (-845) / 4104);
+    // long double k_4 = h * f(t_i + h, y_i + k_0 * 439 / 216 + k_1 * (-8) + k_2 * 3680 / 513 + k_3 * (-845) / 4104);
 
     Vec vec41 = vecInitZerosA(l);
     vecScale(439.0 / 216.0, k_0, &vec41);
@@ -75,7 +75,7 @@ rk45 rkf45_calculator(double h, double t_i, Vec y_i){
 
 
 
-    // double k_5 = h * f(t_i + h * 1 / 2, y_i + k_0 * (-8) / 27 + k_1 * 2 + k_2 * (-3544) / 2565 + k_3 * 1859 / 4104 + k_4 * (-11) / 40);
+    // long double k_5 = h * f(t_i + h * 1 / 2, y_i + k_0 * (-8) / 27 + k_1 * 2 + k_2 * (-3544) / 2565 + k_3 * 1859 / 4104 + k_4 * (-11) / 40);
 
     Vec vec51 = vecInitZerosA(l);
     vecScale(-8.0 / 27.0, k_0, &vec51);
@@ -96,7 +96,7 @@ rk45 rkf45_calculator(double h, double t_i, Vec y_i){
     vecScale(h, f(t_i + h * 0.5, vec51), &k_5);
 
 
-    // double y_order4 = y_i + k_0 * 25 / 216 + k_2 * 1408 / 2565 + k_3 * 2197 / 4104 + k_4 * (-1) / 5;
+    // long double y_order4 = y_i + k_0 * 25 / 216 + k_2 * 1408 / 2565 + k_3 * 2197 / 4104 + k_4 * (-1) / 5;
 
     Vec y_4_1 = vecInitZerosA(l);
     vecScale(25.0 / 216.0, k_0, &y_4_1);
@@ -112,7 +112,7 @@ rk45 rkf45_calculator(double h, double t_i, Vec y_i){
     vecAdd(y_4_1, y_4_4, &y_4_1);
 
 
-    // double y_order5 = y_i + k_0 * 16 / 135 + k_2 * 6656 / 12825 + k_3 * 28561 / 56430 + k_4 * (-9) / 50 + k_5 * 2 / 55;
+    // long double y_order5 = y_i + k_0 * 16 / 135 + k_2 * 6656 / 12825 + k_3 * 28561 / 56430 + k_4 * (-9) / 50 + k_5 * 2 / 55;
 
 
     Vec y_5_1 = vecInitZerosA(l);
@@ -134,7 +134,7 @@ rk45 rkf45_calculator(double h, double t_i, Vec y_i){
     Vec error = vecInitZerosA(l);
     vecSub(y_5_1, y_4_1, &error);
     for(size_t i = 0; i < l; i++){
-        VEC_INDEX(error, i) = fabs(VEC_INDEX(error, i));
+        VEC_INDEX(error, i) = fabsl(VEC_INDEX(error, i));
     }
 
     freeVec(&vec11);
@@ -166,7 +166,7 @@ rk45 rkf45_calculator(double h, double t_i, Vec y_i){
 }
 
 
-void solver(double h, double t_initial, double t_final, Vec y_initial, double TOL, Vec t_res, Mat2d res){
+void solver(long double h, long double t_initial, long double t_final, Vec y_initial, long double TOL, Vec t_res, Mat2d res){
     size_t l = y_initial.len;
     VEC_INDEX(t_res, 0) = t_initial;
     Vec col1 = mat2DCol(res, 0);
@@ -188,7 +188,7 @@ void solver(double h, double t_initial, double t_final, Vec y_initial, double TO
         Vec error = tuple.err;
 
         Vec col = mat2DCol(res, n+1);
-        double errmax = vecMax(error);
+        long double errmax = vecMax(error);
         if(errmax < TOL){
 
             for(size_t i = 0; i < l; i++){
